requirements.txt
Flask==2.0.2
Flask-Cors==3.0.10
Flask-MySQLdb==0.2.0
mysql-connector-python==8.0.26
Flask-Bcrypt==0.7.1
Werkzeug==2.0.2
PyJWT==2.1.0
PyMySQL
dbutils
pymysqlpool
